+++ CHANGELOG.md
@@ -15,6 +15,12 @@ Types of changes:
 - `Security` in case of vulnerabilities.
 -->
 
+[Unreleased]
+
+### Added
+
+- Examples in docstring.
+
 ## [0.1.0]
 
 Initial implementation.
@@ -30,5 +36,6 @@ Initial implementation.
 
 Empty project.
 
+[Unreleased]: https://github.com/FedericoStra/NamedArgs.jl/compare/v0.1.0...HEAD
 [0.1.0]: https://github.com/FedericoStra/NamedArgs.jl/compare/v0.0.0...v0.1.0
 [0.0.0]: https://github.com/FedericoStra/NamedArgs.jl/releases/tag/v0.0.0